chore: eliminate duplicate coercion helpers across Python and React layers
- Add ui/utils/coerce.py and use_cases/_coerce.py as single sources of
  truth for as_str/as_int/as_float and to_int/to_float/to_optional_* helpers
- Remove ~30 copy-pasted coercion functions from 9 GTK4 widget files,
  6 use-case files, the CLI, and the Spotify backend
- Extract sortReleases/parseFocusId/formatSyncSummary/PILL_STYLE shared
  by CollectionPage and WantlistPage into webapp/src/utils/helpers.ts
- Fix deprecated typing.Dict and typing.Optional in image_cache.py and
  settings.py (use built-in dict and X | None instead)
- Add PYTHON_VERSION/NODE_VERSION env vars to 4 workflows to avoid
  duplicated version strings; add explicit permissions: contents: read
  to workflows that lacked a permissions block
