+++ .github/workflows/pilot_validation_windows_macos.yml
@@ -15,6 +15,10 @@ on:
 
 env:
   FORCE_JAVASCRIPT_ACTIONS_TO_NODE24: true
+  PYTHON_VERSION: "3.12"
+
+permissions:
+  contents: read
 
 jobs:
   clean-runner-validation:
@@ -33,7 +37,7 @@ jobs:
       - name: Setup Python
         uses: actions/setup-python@v5
         with:
-          python-version: "3.12"
+          python-version: ${{ env.PYTHON_VERSION }}
 
       - name: Install build tools
         run: |
