+++ src/discogs_player/ui/utils/coerce.py
@@ -1,0 +1,80 @@
+"""Shared type-coercion helpers for the GTK4 widget layer."""
+
+from __future__ import annotations
+
+
+def as_str(value: object | None, fallback: str = "") -> str:
+    if value is None:
+        return fallback
+    result = str(value).strip()
+    return result if result else fallback
+
+
+def as_int(value: object | None, default: int = 0) -> int:
+    if isinstance(value, bool):
+        return int(value)
+    if isinstance(value, (int, float)):
+        return int(value)
+    if isinstance(value, str):
+        text = value.strip()
+        if text:
+            try:
+                return int(text)
+            except ValueError:
+                pass
+    return default
+
+
+def as_float(value: object | None, default: float = 0.0) -> float:
+    if isinstance(value, (int, float)) and not isinstance(value, bool):
+        return float(value)
+    if isinstance(value, str):
+        text = value.strip()
+        if text:
+            try:
+                return float(text)
+            except ValueError:
+                pass
+    return default
+
+
+def as_optional_int(value: object | None) -> int | None:
+    if value is None:
+        return None
+    if isinstance(value, bool):
+        return int(value)
+    if isinstance(value, (int, float)):
+        return int(value)
+    if isinstance(value, str):
+        text = value.strip()
+        if not text:
+            return None
+        try:
+            return int(text)
+        except ValueError:
+            return None
+    try:
+        return int(str(value).strip())
+    except (TypeError, ValueError):
+        return None
+
+
+def as_optional_float(value: object | None) -> float | None:
+    if value is None:
+        return None
+    if isinstance(value, bool):
+        return None
+    if isinstance(value, (int, float)):
+        return float(value)
+    if isinstance(value, str):
+        text = value.strip()
+        if not text:
+            return None
+        try:
+            return float(text)
+        except ValueError:
+            return None
+    try:
+        return float(str(value).strip())
+    except (TypeError, ValueError):
+        return None
