+++ src/discogs_player/ui/widgets/value_dashboard.py
@@ -10,6 +10,9 @@ import gi
 gi.require_version("Gtk", "4.0")
 from gi.repository import GLib, Gtk
 
+from discogs_player.ui.utils.coerce import as_float as _as_float
+from discogs_player.ui.utils.coerce import as_int as _as_int
+from discogs_player.ui.utils.coerce import as_str as _as_str
 from discogs_player.ui.utils.formatting import format_price
 
 _REVEAL_STAGGER_MS = 60
@@ -22,28 +25,6 @@ def _clear_box_children(box: Gtk.Box) -> None:
         next_child = child.get_next_sibling()
         box.remove(child)
         child = next_child
-
-
-def _as_float(value: object | None) -> float:
-    if isinstance(value, (int, float)):
-        return float(value)
-    return 0.0
-
-
-def _as_int(value: object | None) -> int:
-    if isinstance(value, bool):
-        return int(value)
-    if isinstance(value, int):
-        return value
-    if isinstance(value, float):
-        return int(value)
-    return 0
-
-
-def _as_str(value: object | None) -> str:
-    if value is None:
-        return ""
-    return str(value).strip()
 
 
 class ValueDashboard(Gtk.Box):
